requirements.txt
django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-environ==0.11.2
